# gravitee-apim-rest-api/gravitee-apim-rest-api-management-v4/gravitee-apim-rest-api-management-v4-rest/src/main/java/io/gravitee/rest/api/management/v4/rest/resource/connector/EndpointsResource.java
/**
 * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *         http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package io.gravitee.rest.api.management.v4.rest.resource.connector;

import io.gravitee.common.http.MediaType;
import io.gravitee.plugin.core.api.PluginMoreInformation;
import io.gravitee.rest.api.management.v4.rest.mapper.ConnectorPluginMapper;
import io.gravitee.rest.api.management.v4.rest.model.ConnectorPlugin;
import io.gravitee.rest.api.service.v4.EndpointConnectorPluginService;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import java.util.Set;
import javax.inject.Inject;
import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.PathParam;
import javax.ws.rs.Produces;
import javax.ws.rs.container.ResourceContext;
import javax.ws.rs.core.Context;

/**
 * Defines the REST resources to manage endpoints.
 *
 * @author Guillaume LAMIRAND (guillaume.lamirand at graviteesource.com)
 * @author GraviteeSource Team
 */
@Path("/endpoints")
public class EndpointsResource {

    @Context
    private ResourceContext resourceContext;

    @Inject
    private EndpointConnectorPluginService endpointService;

    @GET
    @Produces(MediaType.APPLICATION_JSON)
    public Set<ConnectorPlugin> getEndpoints() {
        return ConnectorPluginMapper.INSTANCE.convertSet(endpointService.findAll());
    }

    @Path("/{endpointId}")
    @GET
    @Produces(MediaType.APPLICATION_JSON)
    public ConnectorPlugin getEndpoint(@PathParam("endpointId") String endpointId) {
        return ConnectorPluginMapper.INSTANCE.convert(endpointService.findById(endpointId));
    }

    @GET
    @Path("/{endpointId}/schema")
    @Produces(MediaType.APPLICATION_JSON)
    public String getEndpointSchema(@PathParam("endpointId") String endpointId) {
        // Check that the endpoint exists
        endpointService.findById(endpointId);

        return endpointService.getSchema(endpointId);
    }

    @GET
    @Path("/{endpointId}/documentation")
    @Produces(MediaType.TEXT_PLAIN)
    public String getEndpointDoc(@PathParam("endpointId") String endpointId) {
        // Check that the endpoint exists
        endpointService.findById(endpointId);

        return endpointService.getDocumentation(endpointId);
    }

    @GET
    @Path("/{endpointId}/moreInformation")
    @ApiResponse(
        responseCode = "200",
        description = "Endpoint more information",
        content = @Content(mediaType = MediaType.APPLICATION_JSON, schema = @Schema(implementation = PluginMoreInformation.class))
    )
    @Produces(MediaType.APPLICATION_JSON)
    public PluginMoreInformation getMoreInformation(@PathParam("endpointId") String endpointId) {
        // Check that the entrypoint exists
        endpointService.findById(endpointId);

        return endpointService.getMoreInformation(endpointId);
    }

    @GET
    @Path("/{endpointId}/sharedConfigurationSchema")
    @Produces(MediaType.APPLICATION_JSON)
    public String getSharedConfigurationSchema(@PathParam("endpointId") String endpointId) {
        // Check that the entrypoint exists
        endpointService.findById(endpointId);

        return endpointService.getSharedConfigurationSchema(endpointId);
    }
}